Add CSVSaver, Add NetworkSaver
Actually, it can't compile, because i didn't finish everything:
-Convert string to float
-Make the GenerationHandler load the networks contained in the csv files under some conditions

+++ Window.h
@@ -11,7 +11,7 @@ public:
 	void draw_triangle(float base, float side1, float side2, Color color);
 	void draw_line(Pos origin, Pos end, Color color);
 	void draw_square(Pos pos, float side, Color color);
-	void draw_circle(Pos center, float radius, Color color);
+	void draw_circle(Pos center, double radius, Color color);
 	//void drawSequence(List* vertices);
 	void update();
 	GLFWwindow* window;
